+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
+- FF-1429 - Updated Microsoft.NET.Test.Sdk to 16.7.0
 - FF-1429 - Updated Microsoft.CodeAnalysis.CSharp.Workspaces to 3.7.0
 - FF-1429 - Updated FunFair.Test.Common to 1.10.1.439
 - FF-1429 - Updated xunit.runner.visualstudio to 2.4.3
@@ -113,6 +114,7 @@ Releases that have at least been deployed to staging, BUT NOT necessarily releas
 ## [1.0.0] - 2019-12-18
 ### Added
 - Banned DateTime.Now, DateTime.UtcNow, DateTime.Today, DateTimeOffset.Now and DateTimeOffset.UtcNow pointing to use DateTimeSource instead
+