+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
+- FF-1429 - Updated FunFair.Test.Common to 1.11.2.471
 - FF-1429 - Updated FunFair.Test.Common to 1.11.1.466
 - FF-1429 - Updated FunFair.Test.Common to 1.11.0.461
 ### Removed