+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
+- FF-2802 - Prohibit Received(0) in tests
 - FF-1429 - Updated FunFair.Test.Common to 1.11.2.471
 - FF-1429 - Updated Microsoft.NET.Test.Sdk to 16.7.1
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.12.0.21095